backproject/requirements.txt
Django==3.2.15
djangorestframework==3.14.0
InstructorEmbedding==1.0.1
numpy==1.24.3
pandas==1.5.3
replicate==0.13.0
scikit_learn==1.2.2
torch==1.13.1
